+++ core/client/app/styles/layouts/content.css
@@ -1,12 +1,4 @@
 /* Content /ghost/
-/* ---------------------------------------------------------- */
-
-.content-view-container {
-    position: relative;
-    flex-grow: 1;
-    display: flex;
-    flex-direction: column;
-}
 
 /* Show/Hide on Mobile // TODO: What the fuck does that mean?
 /* ---------------------------------------------------------- */
