+++ core/client/app/templates/posts.hbs
@@ -1,3 +1,4 @@
+<section class="gh-view content-view-container">
 <header class="view-header">
     <h2 class="view-title">Content</h2>
     <section class="view-actions">
@@ -40,3 +41,4 @@
         {{outlet}}
     </section>
 </div>
+</section>
