--- core/client/app/views/posts/post.js
@@ -1,5 +1,0 @@
-import MobileContentView from 'ghost/views/mobile/content-view';
-
-var PostsPostView = MobileContentView.extend();
-
-export default PostsPostView;
