+++ core/client/app/views/settings/content-base.js
@@ -1,13 +1,8 @@
-import MobileContentView from 'ghost/views/mobile/content-view';
-/**
- * All settings views other than the index should inherit from this base class.
- * It ensures that the correct screen is showing when a mobile user navigates
- * to a `settings.someRouteThatIsntIndex` route.
- */
+import Ember from 'ember';
 
-var SettingsContentBaseView = MobileContentView.extend({
+var SettingsView = Ember.View.extend({
     tagName: 'section',
-    classNames: ['gh-view', 'js-settings-content']
+    classNames: ['gh-view']
 });
 
-export default SettingsContentBaseView;
+export default SettingsView;
